+++ src/utils/ActionOptions.ts
@@ -1,6 +1,6 @@
-import * as core from '@actions/core';
 import * as path from 'path';
 import { FailOn, FailOnEnum } from './FailOn';
+import { getInputSafe } from './getInput';
 
 /**
  * Contains all the options of the action
@@ -14,14 +14,14 @@ export class ActionOptions {
   public readonly format: boolean;
   public readonly lineLength: number | null;
   constructor() {
-    this.failOn = FailOn.fromInput(core.getInput('fail-on', { required: true }));
-    this.workingDirectory = path.resolve(process.env.GITHUB_WORKSPACE!, core.getInput('working-directory', { required: true }));
-    this.token = core.getInput('token', { required: true });
-    this.checkRenamedFiles = core.getInput('check-renamed-files', { required: true }) === 'true';
-    this.emojis = core.getInput('emojis', { required: true }) === 'true';
-    this.format = core.getInput('format', { required: true }) === 'true';
+    this.failOn = FailOn.fromInput(getInputSafe('fail-on') || 'error');
+    this.token = getInputSafe('token', { required: true });
+    this.workingDirectory = path.resolve(process.env.GITHUB_WORKSPACE!, getInputSafe('working-directory') ?? './');
+    this.checkRenamedFiles = getInputSafe('check-renamed-files') === 'true';
+    this.emojis = (getInputSafe('emojis') || 'true') === 'true';
+    this.format = (getInputSafe('format') || 'true') === 'true';
     try {
-      this.lineLength = parseInt(core.getInput('line-length', { required: true }));
+      this.lineLength = parseInt(getInputSafe('line-length'));
     } catch (_) {
       this.lineLength = null;
     }