+++ src/utils/ModifiedFiles.ts
@@ -2,8 +2,9 @@ import * as core from '@actions/core';
 import * as github from '@actions/github';
 import { context } from '@actions/github/lib/utils';
 import path from 'path';
-import { EventName } from '../actions/github/EventName';
 import { actionOptions } from './ActionOptions';
+import type { EventName } from '../Actions/Github/EventName';
+import { getInputSafe } from './getInput';
 
 interface FileLinesInterface {
   start: number;
@@ -219,7 +220,7 @@ export class ModifiedFiles {
     }
 
     /// Github client from API token
-    const client = github.getOctokit(core.getInput('token', { required: true }));
+    const client = github.getOctokit(getInputSafe('token', { required: true }));
 
     const response = await client.repos.compareCommits({
       base,